+++ app/User.php
@@ -60,6 +60,11 @@ class User extends Model
         return $this->hasMany('App\Message');
     }
 
+    public function MAMEarliests()
+    {
+        return $this->hasMany('App\MAMEarliest');
+    }
+
     public function openChats()
     {
         return $this->hasMany('App\OpenChat');